Versao 6.6.2

+++ frontend/src/translate/languages/pt.js
@@ -160,8 +160,7 @@ const messages = {
           deleteTitle: "Deletar",
           deleteMessage: "Você tem certeza? Essa ação não pode ser revertida.",
           disconnectTitle: "Desconectar",
-          disconnectMessage:
-            "Tem certeza? Você precisará ler o QR Code novamente.",
+          disconnectMessage: "Tem certeza? Você precisará ler o QR Code novamente.",
         },
         buttons: {
           add: "Adicionar WhatsApp",
@@ -200,7 +199,7 @@ const messages = {
         },
         table: {
           name: "Nome",
-		  number: "Número",
+		      number: "Número",
           status: "Status",
           lastUpdate: "Última atualização",
           default: "Padrão",
@@ -213,29 +212,31 @@ const messages = {
           add: "Adicionar WhatsApp",
           edit: "Editar WhatsApp",
         },		
-		 tabs: {
+		    tabs: {
           general: "Geral",
           messages: "Mensagens",
           assessments: "Avaliações",
           integrations: "Integrações",
           schedules: "Horário de expediente",
-		 },
+		    },
         form: {
           name: "Nome",
           default: "Padrão",
           sendIdQueue: "Fila",
           timeSendQueue: "Redirecionar para fila em X minutos",
           queueRedirection: "Redirecionamento de Fila",
-		  outOfHoursMessage: "Mensagem de fora de expediente",
+		      outOfHoursMessage: "Mensagem de fora de expediente",
           queueRedirectionDesc: "Selecione uma fila para os contatos que não possuem fila serem redirecionados",
           prompt: "Prompt",
           greetingMedia: "Mídia de Saudação (Opcional)",
           attachImage: "Anexar Imagem",
           transferAfterMinutes: "Transferir após x (minutos)",
           transferQueue: "Fila de Transferência",
+          // maxUseBotQueues: "Enviar bot x vezes",
+          // timeUseBotQueues: "Intervalo em minutos entre envio de bot",
           expiresTicket: "Encerrar chats abertos após x minutos",
           expiresInactiveMessage: "Mensagem de encerramento por inatividade",
-		  greetingMessage: "Mensagem de saudação",
+		      greetingMessage: "Mensagem de saudação",
           complationMessage: "Mensagem de conclusão",
         },
         buttons: {
@@ -254,8 +255,8 @@ const messages = {
           deleted: "Contato excluído com sucesso!",
 		      deletedAll: "Todos contatos excluídos com sucesso!",
         },
-		  searchPlaceholder: "Pesquisar...",
-          confirmationModal: {
+		    searchPlaceholder: "Pesquisar...",
+        confirmationModal: {
           deleteTitle: "Deletar",
           deleteAllTitle: "Deletar Todos",
           importTitle: "Importar contatos",
@@ -265,7 +266,7 @@ const messages = {
         },
         buttons: {
           import: "Importar Contatos",
-		  importSheet: "Import. Excel",
+		      importSheet: "Import. Excel",
           add: "Adicionar Contato",
           export: "Exportar Contatos",
           delete: "Excluir Todos Contatos",
@@ -319,14 +320,14 @@ const messages = {
           editSuccess: "Integração editada com sucesso.",
         },
       },
-	  sideMenu: {
-            name: "Menu Lateral Inicial",
-            note: "Se habilitado, o menu lateral irá iniciar fechado",
-            options: {
-              enabled: "Aberto",
-              disabled: "Fechado",
-            },
-          },
+	    sideMenu: {
+        name: "Menu Lateral Inicial",
+        note: "Se habilitado, o menu lateral irá iniciar fechado",
+        options: {
+          enabled: "Aberto",
+          disabled: "Fechado",
+        },
+      },
       promptModal: {
         form: {
           name: "Nome",
@@ -451,7 +452,7 @@ const messages = {
           queueData: "Dados da Fila",
           serviceHours: "Horários de Atendimento",
         },
-	toasts: {
+	      toasts: {
           deleted: "Registro excluído",
         },      
         confirmationModal: {
@@ -474,9 +475,9 @@ const messages = {
           okAdd: "Adicionar",
           okEdit: "Salvar",
           cancel: "Cancelar",
-		  attach: "Anexar Arquivo",
+		      attach: "Anexar Arquivo",
         },
-		 serviceHours: {
+		    serviceHours: {
           dayWeek: "Dia da semana",
           startTimeA: "Hora Inicial - 1",
           endTimeA: "Hora Final - 1",
@@ -678,9 +679,9 @@ const messages = {
         buttons: {
           accept: "Aceitar",
           closed: "Finalizar",
-		  transfer: "Transferir",
+		      transfer: "Transferir",
           reopen: "Reabrir",
-		  exportAsPDF: "Exportar em PDF"
+		      exportAsPDF: "Exportar em PDF"
         },
       },
       newTicketModal: {
@@ -728,7 +729,7 @@ const messages = {
           financeiro: "Financeiro",
           files: "Lista de arquivos",
           prompts: "Open.Ai",
-		  reports: "Relatórios",
+		      reports: "Relatórios",
           queueIntegration: "Integrações",
           languageSettings: "Configuração de Idioma",
         },
@@ -761,7 +762,7 @@ const messages = {
             "Você tem certeza? Essa ação não pode ser revertida! e será removida das filas e conexões vinculadas",
         },
       },
-	  reports: {
+	    reports: {
         title: "Relatórios de Atendimentos",
         labels: {
           contact: "Contato",
@@ -1221,7 +1222,7 @@ const messages = {
       queues: {
         title: "Filas & Chatbot",
         table: {
-		  id:"ID",
+		      id:"ID",
           name: "Nome",
           color: "Cor",
           greeting: "Mensagem de saudação",
@@ -1243,14 +1244,14 @@ const messages = {
       users: {
         title: "Usuários",
         table: {
-		  id: "ID",
+		      id: "ID",
           name: "Nome",
-		  status: "Status",
+		      status: "Status",
           email: "Email",
           profile: "Perfil",
           actions: "Ações",
         },
-		status: {
+		    status: {
           online: "Usuários online",
           offline: "Usuários offline",
         },
@@ -1319,8 +1320,8 @@ const messages = {
         confirmationModal: {
           deleteTitle: "Você tem certeza que quer excluir esta Tag?",
           deleteMessage: "Esta ação não pode ser revertida.",
-		  deleteAllMessage: "Tem certeza que deseja deletar todas as Tags?",
-		  deleteAllTitle: "Deletar Todos",
+		      deleteAllMessage: "Tem certeza que deseja deletar todas as Tags?",
+		      deleteAllTitle: "Deletar Todos",
         },
         table: {
           name: "Nome",
@@ -1330,10 +1331,10 @@ const messages = {
         },
         buttons: {
           add: "Nova Tag",
-		  deleteAll: "Deletar Todas",
+		      deleteAll: "Deletar Todas",
         },
         toasts: {
-		  deletedAll: "Todas Tags excluídas com sucesso!",
+		      deletedAll: "Todas Tags excluídas com sucesso!",
           deleted: "Tag excluído com sucesso.",
         },
       },
@@ -1594,7 +1595,7 @@ const messages = {
         },
         confirmationModal: {
           title: "Deletar o ticket",
-		  titleFrom: "do contato ",
+		      titleFrom: "do contato ",
           message:
             "Atenção! Todas as mensagens relacionadas ao ticket serão perdidas.",
         },
@@ -1611,12 +1612,12 @@ const messages = {
       },
       messageOptionsMenu: {
         delete: "Deletar",
-		reactionSuccess: "Reação add",
+		    reactionSuccess: "Reação add",
         reply: "Responder",
-		edit: 'Editar Mensagem',
-		forward: "Encaminhar",
+		    edit: 'Editar Mensagem',
+		    forward: "Encaminhar",
         toForward: "Encaminhar",
-		react: "Reagir",
+		    react: "Reagir",
         confirmationModal: {
           title: "Apagar mensagem?",
           message: "Esta ação não pode ser revertida.",
